+++ R/plot_Ages.R
@@ -6,7 +6,7 @@
 #'
 #' @param object [list] (**required**): Output as created by functions like [AgeC14_Computation].
 #'
-#' @param sample_names [character] (optional): alternative samples names used for the plotting, this can
+#' @param sample_names [character] (optional): alternative sample names used for the plotting, this can
 #' also be used to resort the samples in a specific order. If the length of the provided [character] vector
 #' is shorter than the real number of samples, the names are recycled.
 #'
@@ -17,7 +17,7 @@
 #' @return
 #' The function returns a plot and the [data.frame] used to display the data
 #'
-#' @section Function version: 0.1.0
+#' @section Function version: 0.1.1
 #'
 #' @author Sebastian Kreutzer, IRAMAT-CRP2A, UMR 5060, CNRS - Université Bordeaux Montaigne (France), based on code
 #' written by Claire Christophe
@@ -106,7 +106,7 @@ plot_Ages <- function(
 
   ##PLOTTING
   ##adjust par
-  par(mfrow=c(1,1),las = 1,oma=c(2,5,0.5,0.5), cex = plot_settings$cex)
+  par(mfrow=c(1,1),las = 1, oma = c(2,5,0.5,0.5), cex = plot_settings$cex)
 
   ##open plot area
   plot(x = NA,
@@ -122,7 +122,7 @@ plot_Ages <- function(
   ##add y-axis
   axis(
     side = 2,
-    at = as.factor(df[["SAMPLE"]]),
+    at = df[["AT"]],
     labels = if(!is.null(sample_names)){
      df[["ALT_SAMPLE_NAME"]]
     }else{