+++ NEWS.md
@@ -1,6 +1,7 @@
 ## BayLum 0.1.4 (Release date: 2018-09-19)
 
 * `Generate_DataFile_MG()` now works only on preselected records; the limitation of OSL only was removed.
+* `plot_Ages()` the sorting of the samples in the graph did not work as expected.
 
 ## BayLum 0.1.3 (Release date: 2018-09-14)